requirements.txt
Flask
python-dotenv
google-generativeai
markdown